url bugfixed

+++ app/Http/Controllers/AdminController.php
@@ -10,19 +10,29 @@ use App\Models\Order;
 use App\Models\Reservation;
 use Illuminate\Support\Facades\Auth;
 
+// function isLoggedIn(){
+//     dd(Auth::user());
+//     if (Auth::user()->usertype == '0'){
+//         return redirect('redirects');
+//     }
+// }
 
 class AdminController extends Controller
 {
-
     //user part
     public function user()
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = user::all();
         return view("admin.users", compact("data"));
-       
     }
     public function deleteuser($id)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = user::find($id);
         $data->delete();
         return redirect()->back();
@@ -31,12 +41,17 @@ class AdminController extends Controller
     // FoodMenu Part
     public function foodmenu()
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = foods::all();
         return view("admin.foodmenu", compact("data"));
     }
     public function uploadfoodmenu(Request $request)
     {
-
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = new foods;
 
         $image = $request->image;
@@ -57,6 +72,9 @@ class AdminController extends Controller
 
     public function deletemenu($id)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = foods::find($id);
         $data->delete();
         return redirect()->back();
@@ -64,12 +82,18 @@ class AdminController extends Controller
 
     public function editfoodmenu($id)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = foods::find($id);
         return view("admin.editmenu", compact("data"));
     }
 
     public function update(Request $request, $id)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = foods::find($id);
 
         $image = $request->image;
@@ -97,36 +121,40 @@ class AdminController extends Controller
 
     public function reservation(Request $request)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $usertype = Auth::user()->usertype;
 
         if ($usertype == '1') {
 
             return view('admin.adminhome');
         } else {
-        $data = new Reservation;
+            $data = new Reservation;
 
-        $data->name = $request->name;
-        $data->email = $request->email;
-        $data->phone = $request->phone;
-        $data->guest = $request->guest;
-        $data->date = $request->date;
-        $data->time = $request->time;
-        $data->message = $request->message;
+            $data->name = $request->name;
+            $data->email = $request->email;
+            $data->phone = $request->phone;
+            $data->guest = $request->guest;
+            $data->date = $request->date;
+            $data->time = $request->time;
+            $data->message = $request->message;
 
-        $data->save();
+            $data->save();
 
-        return redirect()->back();
+            return redirect()->back();
         }
     }
 
     public function viewreservation()
     {
-        if(Auth::id()){
-        $data = reservation::all();
-        return view("admin.reservation", compact("data"));
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
         }
-        else
-        {
+        if (Auth::id()) {
+            $data = reservation::all();
+            return view("admin.reservation", compact("data"));
+        } else {
             return redirect('login');
         }
     }
@@ -136,18 +164,22 @@ class AdminController extends Controller
 
     public function viewchefs()
     {
-         if(Auth::id()){
-        $data = chefs::all();
-        return view("admin.chefs", compact("data"));
-    }
-    else
-    {
-        return redirect('login');
-    }
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
+        if (Auth::id()) {
+            $data = chefs::all();
+            return view("admin.chefs", compact("data"));
+        } else {
+            return redirect('login');
+        }
     }
 
     public function chefinfoupload(Request $request)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = new Chefs;
 
         $image = $request->image;
@@ -168,6 +200,9 @@ class AdminController extends Controller
 
     public function deletchefinfo($id)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = chefs::find($id);
         $data->delete();
         return redirect()->back();
@@ -175,11 +210,17 @@ class AdminController extends Controller
 
     public function updatechefinfo($id)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = chefs::find($id);
         return view("admin.editchefinfo", compact("data"));
     }
     public function editchefinfo(Request $request, $id)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = chefs::find($id);
 
         $image = $request->image;
@@ -200,16 +241,21 @@ class AdminController extends Controller
 
     public function orders()
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $data = Order::all();
 
-        return view('admin.orders',compact("data"));
+        return view('admin.orders', compact("data"));
     }
 
     public function search(Request $request)
     {
+        if (Auth::user()->usertype == '0') {
+            return redirect('redirects');
+        }
         $search = $request->search;
-        $data = Order::where('username','Like','%'.$search.'%')->orWhere('foodname','Like','%'.$search.'%')->orWhere('address','Like','%'.$search.'%')->get();
-        return view('admin.orders',compact("data"));
+        $data = Order::where('username', 'Like', '%' . $search . '%')->orWhere('foodname', 'Like', '%' . $search . '%')->orWhere('address', 'Like', '%' . $search . '%')->get();
+        return view('admin.orders', compact("data"));
     }
-
 }
